% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/it_randomize.R
\name{it_randomize}
\alias{it_randomize}
\title{Randomize a matrix using the IT algorithm}
\usage{
it_randomize(matr, n_iter = 10000)
}
\arguments{
\item{matr}{A community matrix to be randomized.}

\item{n_iter}{Number of iterations to attempt to randomize the matrix.}
}
\value{
A randomized version of `matr` where the row
and column totals match that of `matr`, and the values
for each element were sampled proportional to the
corresponding element value in `matr`
}
\description{
More info needed here
}
\details{
Follows alogrithm as described in Ulrich and Gotelli 2010.
As per this source "IT assigns individuals randomly to matrix
cells with probabilities proportional to observed row and
column abundance totals until, for each row and column, total
abundances are reached."
}
