% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_ratio_calculator.R
\name{u_ratio_calculator}
\alias{u_ratio_calculator}
\title{Calculate the U-ratio for a Community Matrix}
\usage{
u_ratio_calculator(matr)
}
\arguments{
\item{matr}{A community matrix with \emph{sites} as columns
and \emph{species} as rows}
}
\value{
U-ratio
}
\description{
\code{u_ratio_calculator} returns the U-ratio
value for a community matrix, following the description of
this metric in Ulrich and Gotelli 2010 and Schroeder et al.
2014, PLoS ONE.
}
\details{
As described in Schroeder et al. 2014, " U-ratio = V/W ; metric of co-occurrence, which compares the variance of row totals V with the sum of the column variances W. Low values of U indicate negative covariation in abundance between species. If species are segregated, we expect the U-ratio to be smaller than expected by chance.
}
